Refactor JDA and role management to JdaService
Moved JDA instance and role management from AsyncTI4DiscordBot to a new JdaService class under ti4.spring.jda. Updated all references throughout the codebase to use JdaService for accessing JDA, guilds, and role sets. This improves separation of concerns and prepares for better Spring integration.
