+++ src/main/java/ti4/commands/async/ShowHeroes.java
@@ -6,7 +6,6 @@ import net.dv8tion.jda.api.entities.Role;
 import net.dv8tion.jda.api.events.interaction.GenericInteractionCreateEvent;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import org.apache.commons.lang3.StringUtils;
-import ti4.AsyncTI4DiscordBot;
 import ti4.commands.Subcommand;
 import ti4.helpers.TIGLHelper;
 import ti4.helpers.TIGLHelper.TIGLRank;
@@ -14,6 +13,7 @@ import ti4.message.MessageHelper;
 import ti4.service.emoji.FactionEmojis;
 import ti4.service.emoji.LeaderEmojis;
 import ti4.service.emoji.TI4Emoji;
+import ti4.spring.jda.JdaService;
 
 class ShowHeroes extends Subcommand {
 
@@ -36,7 +36,7 @@ class ShowHeroes extends Subcommand {
             String faction = StringUtils.substringAfter(rank.toString(), "_");
             TI4Emoji factionIcon = FactionEmojis.getFactionIcon(faction);
             TI4Emoji heroEmoji = LeaderEmojis.getLeaderEmoji(faction + "hero");
-            List<Member> members = AsyncTI4DiscordBot.guildPrimary.getMembersWithRoles(role);
+            List<Member> members = JdaService.guildPrimary.getMembersWithRoles(role);
 
             sb.append("> ").append(factionIcon);
             for (Member member : members) {
