+++ src/main/java/ti4/commands/bothelper/BothelperCommand.java
@@ -4,11 +4,11 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
-import ti4.AsyncTI4DiscordBot;
 import ti4.commands.CommandHelper;
 import ti4.commands.ParentCommand;
 import ti4.commands.Subcommand;
 import ti4.helpers.Constants;
+import ti4.spring.jda.JdaService;
 
 public class BothelperCommand implements ParentCommand {
 
@@ -35,8 +35,7 @@ public class BothelperCommand implements ParentCommand {
 
     @Override
     public boolean accept(SlashCommandInteractionEvent event) {
-        return ParentCommand.super.accept(event)
-                && CommandHelper.acceptIfHasRoles(event, AsyncTI4DiscordBot.bothelperRoles);
+        return ParentCommand.super.accept(event) && CommandHelper.acceptIfHasRoles(event, JdaService.bothelperRoles);
     }
 
     @Override
