+++ src/main/java/ti4/commands/bothelper/JazzCommand.java
@@ -6,7 +6,6 @@ import java.util.List;
 import net.dv8tion.jda.api.components.buttons.Button;
 import net.dv8tion.jda.api.events.interaction.GenericInteractionCreateEvent;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
-import ti4.AsyncTI4DiscordBot;
 import ti4.buttons.Buttons;
 import ti4.commands.Subcommand;
 import ti4.helpers.Constants;
@@ -15,6 +14,7 @@ import ti4.listeners.annotations.ButtonHandler;
 import ti4.message.MessageHelper;
 import ti4.message.logging.BotLogger;
 import ti4.service.emoji.MiscEmojis;
+import ti4.spring.jda.JdaService;
 
 class JazzCommand extends Subcommand {
 
@@ -58,13 +58,13 @@ class JazzCommand extends Subcommand {
 
     @ButtonHandler("jazzButton")
     private static void jazzButton() {
-        AsyncTI4DiscordBot.jda
+        JdaService.jda
                 .getGuildById("847560709730730064")
                 .getTextChannelById("1352824638354231439")
                 .sendMessage("```fix\nBorgJedi used /search my_titles player: @Mentak\n```")
                 .queue();
 
-        AsyncTI4DiscordBot.jda
+        JdaService.jda
                 .getGuildById("847560709730730064")
                 .getTextChannelById("1352824638354231439")
                 .sendMessage("**__Mentak's Titles__**\n` 1.`**You Made Me Mad** x5 (g15, g15, g15, g15, g15)")
