+++ src/main/java/ti4/commands/bothelper/ServerGameStats.java
@@ -11,12 +11,12 @@ import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEve
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
-import ti4.AsyncTI4DiscordBot;
 import ti4.commands.Subcommand;
 import ti4.helpers.Constants;
 import ti4.map.persistence.GameManager;
 import ti4.map.persistence.ManagedGame;
 import ti4.message.MessageHelper;
+import ti4.spring.jda.JdaService;
 
 class ServerGameStats extends Subcommand {
 
@@ -37,7 +37,7 @@ class ServerGameStats extends Subcommand {
         int hostedGames = 0;
         int roomForGames = 0;
 
-        List<Guild> guilds = AsyncTI4DiscordBot.guilds.stream()
+        List<Guild> guilds = JdaService.guilds.stream()
                 .filter(g -> !skipGuilds.contains(g.getId()))
                 .sorted(Comparator.comparing(Guild::getIdLong)) // Sort by creation date
                 .toList();
