+++ src/main/java/ti4/commands/game/Swap.java
@@ -6,7 +6,6 @@ import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEve
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
-import ti4.AsyncTI4DiscordBot;
 import ti4.commands.CommandHelper;
 import ti4.commands.GameStateSubcommand;
 import ti4.helpers.Constants;
@@ -14,6 +13,7 @@ import ti4.map.Game;
 import ti4.map.Player;
 import ti4.message.MessageHelper;
 import ti4.service.game.SwapFactionService;
+import ti4.spring.jda.JdaService;
 
 class Swap extends GameStateSubcommand {
 
@@ -31,7 +31,7 @@ class Swap extends GameStateSubcommand {
         User callerUser = event.getUser();
         Game game = getGame();
         Collection<Player> players = game.getPlayers().values();
-        boolean isAdmin = CommandHelper.hasRole(event, AsyncTI4DiscordBot.adminRoles);
+        boolean isAdmin = CommandHelper.hasRole(event, JdaService.adminRoles);
         if (players.stream().noneMatch(player -> player.getUserID().equals(callerUser.getId())) && !isAdmin) {
             MessageHelper.sendMessageToChannel(event.getChannel(), "Only game players can swap with a player.");
             return;
