+++ src/main/java/ti4/listeners/AutoCompleteListener.java
@@ -3,9 +3,9 @@ package ti4.listeners;
 import javax.annotation.Nonnull;
 import net.dv8tion.jda.api.events.interaction.command.CommandAutoCompleteInteractionEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
-import ti4.AsyncTI4DiscordBot;
 import ti4.autocomplete.AutoCompleteProvider;
 import ti4.executors.ExecutorServiceManager;
+import ti4.spring.jda.JdaService;
 
 public class AutoCompleteListener extends ListenerAdapter {
 
@@ -13,7 +13,7 @@ public class AutoCompleteListener extends ListenerAdapter {
 
     @Override
     public void onCommandAutoCompleteInteraction(@Nonnull CommandAutoCompleteInteractionEvent event) {
-        if (!AsyncTI4DiscordBot.isReadyToReceiveCommands()
+        if (!JdaService.isReadyToReceiveCommands()
                 && !"developer setting".equals(event.getInteraction().getFullCommandName())) {
             event.replyChoice("Please try again in a moment. The bot is not ready to serve AutoComplete.", 0)
                     .queue();
