+++ src/main/java/ti4/listeners/DeletionListener.java
@@ -8,31 +8,31 @@ import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.channel.concrete.TextChannel;
 import net.dv8tion.jda.api.events.message.MessageDeleteEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
-import ti4.AsyncTI4DiscordBot;
 import ti4.message.MessageHelper;
 import ti4.message.logging.BotLogger;
+import ti4.spring.jda.JdaService;
 
 public class DeletionListener extends ListenerAdapter {
 
     @Override
     public void onMessageDelete(@Nonnull MessageDeleteEvent event) {
-        if (!validateEvent(event)) return;
+        if (!validateEvent(event)) {}
         // ExecutorServiceManager.runAsync("DeletionListener task", () -> handleMessageDelete(event));
         // reinstate this when it works/someone has time to make it work.
     }
 
     private static boolean validateEvent(MessageDeleteEvent event) {
-        if (!AsyncTI4DiscordBot.isReadyToReceiveCommands()) {
+        if (!JdaService.isReadyToReceiveCommands()) {
             return false;
         }
         String eventGuild = event.getGuild().getId();
-        return AsyncTI4DiscordBot.isValidGuild(eventGuild);
+        return JdaService.isValidGuild(eventGuild);
     }
 
     private void handleMessageDelete(MessageDeleteEvent event) {
         try {
             TextChannel deletionLogChannel =
-                    AsyncTI4DiscordBot.guildPrimary.getTextChannelsByName("deletion-log", true).stream()
+                    JdaService.guildPrimary.getTextChannelsByName("deletion-log", true).stream()
                             .findFirst()
                             .orElse(null);
             if (deletionLogChannel == null) return;
