+++ src/main/java/ti4/map/Game.java
@@ -1,7 +1,7 @@
 package ti4.map;
 
-import static java.util.function.Predicate.*;
-import static org.apache.commons.collections4.CollectionUtils.*;
+import static java.util.function.Predicate.not;
+import static org.apache.commons.collections4.CollectionUtils.isNotEmpty;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
@@ -46,7 +46,6 @@ import net.dv8tion.jda.internal.utils.tuple.ImmutablePair;
 import net.dv8tion.jda.internal.utils.tuple.Pair;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
-import ti4.AsyncTI4DiscordBot;
 import ti4.commands.planet.PlanetRemove;
 import ti4.draft.BagDraft;
 import ti4.draft.DraftItem;
@@ -100,6 +99,7 @@ import ti4.service.emoji.SourceEmojis;
 import ti4.service.leader.CommanderUnlockCheckService;
 import ti4.service.milty.MiltyDraftManager;
 import ti4.service.option.FOWOptionService.FOWOption;
+import ti4.spring.jda.JdaService;
 
 public class Game extends GameProperties {
 
@@ -200,7 +200,7 @@ public class Game extends GameProperties {
     private MiltyDraftManager miltyDraftManager;
 
     @Getter
-    private Expeditions expeditions = new Expeditions(this);
+    private final Expeditions expeditions = new Expeditions(this);
 
     @Setter
     @Getter
@@ -769,10 +769,10 @@ public class Game extends GameProperties {
     @Nullable
     public TextChannel getTableTalkChannel() {
         try {
-            return AsyncTI4DiscordBot.jda.getTextChannelById(getTableTalkChannelID());
+            return JdaService.jda.getTextChannelById(getTableTalkChannelID());
         } catch (Exception e) {
             TextChannel tableTalkChannel;
-            List<TextChannel> gameChannels = AsyncTI4DiscordBot.jda.getTextChannels().stream()
+            List<TextChannel> gameChannels = JdaService.jda.getTextChannels().stream()
                     .filter(c -> c.getName().startsWith(getName()))
                     .filter(not(c -> c.getName().contains(Constants.ACTIONS_CHANNEL_SUFFIX)))
                     .toList();
@@ -788,10 +788,10 @@ public class Game extends GameProperties {
     @JsonIgnore
     public TextChannel getMainGameChannel() {
         try {
-            return AsyncTI4DiscordBot.jda.getTextChannelById(getMainChannelID());
+            return JdaService.jda.getTextChannelById(getMainChannelID());
         } catch (Exception e) {
             List<TextChannel> gameChannels =
-                    AsyncTI4DiscordBot.jda.getTextChannelsByName(getName() + Constants.ACTIONS_CHANNEL_SUFFIX, true);
+                    JdaService.jda.getTextChannelsByName(getName() + Constants.ACTIONS_CHANNEL_SUFFIX, true);
             if (gameChannels.size() == 1) {
                 TextChannel mainGameChannel = gameChannels.getFirst();
                 setMainChannelID(mainGameChannel.getId());
@@ -805,7 +805,7 @@ public class Game extends GameProperties {
     @JsonIgnore
     public TextChannel getSavedChannel() {
         try {
-            return AsyncTI4DiscordBot.jda.getTextChannelById(getSavedChannelID());
+            return JdaService.jda.getTextChannelById(getSavedChannelID());
         } catch (Exception e) {
             return getMainGameChannel();
         }
@@ -824,7 +824,7 @@ public class Game extends GameProperties {
 
         // FIND BY ID
         if (StringUtils.isNumeric(getBotMapUpdatesThreadID())) {
-            ThreadChannel threadChannel = AsyncTI4DiscordBot.jda.getThreadChannelById(getBotMapUpdatesThreadID());
+            ThreadChannel threadChannel = JdaService.jda.getThreadChannelById(getBotMapUpdatesThreadID());
             if (threadChannel != null) {
                 return threadChannel;
             }
@@ -832,7 +832,7 @@ public class Game extends GameProperties {
 
         // FIND BY NAME
         List<ThreadChannel> botChannels =
-                AsyncTI4DiscordBot.jda.getThreadChannelsByName(getName() + Constants.BOT_CHANNEL_SUFFIX, true);
+                JdaService.jda.getThreadChannelsByName(getName() + Constants.BOT_CHANNEL_SUFFIX, true);
         if (botChannels.size() == 1) {
             return botChannels.getFirst();
         } else if (botChannels.size() > 1) {
@@ -862,7 +862,7 @@ public class Game extends GameProperties {
 
     public ThreadChannel getLaunchPostThread() {
         if (StringUtils.isNumeric(getLaunchPostThreadID())) {
-            return AsyncTI4DiscordBot.guildPrimary.getThreadChannelById(getLaunchPostThreadID());
+            return JdaService.guildPrimary.getThreadChannelById(getLaunchPostThreadID());
         }
         return null;
     }
@@ -3234,7 +3234,7 @@ public class Game extends GameProperties {
         setStrategyCardSet(deckSettings.getStratCards().getChosenKey());
 
         // Setup peakable objectives
-        if (getPublicObjectives1Peakable().size() != 4) {
+        if (publicObjectives1Peakable.size() != 4) {
             setUpPeakableObjectives(miltySettings.getGameSettings().getStage1s().getVal(), 1);
             setUpPeakableObjectives(miltySettings.getGameSettings().getStage2s().getVal(), 2);
         }
@@ -3527,7 +3527,7 @@ public class Game extends GameProperties {
         }
         StringBuilder sb = new StringBuilder(getName()).append(" ");
         for (String playerID : getPlayerIDs()) {
-            User user = AsyncTI4DiscordBot.jda.getUserById(playerID);
+            User user = JdaService.jda.getUserById(playerID);
             if (user != null) sb.append(user.getAsMention()).append(" ");
         }
         return sb.toString();
