+++ src/main/java/ti4/map/Player.java
@@ -36,7 +36,6 @@ import net.dv8tion.jda.api.requests.restaction.ThreadChannelAction;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
-import ti4.AsyncTI4DiscordBot;
 import ti4.buttons.Buttons;
 import ti4.draft.DraftBag;
 import ti4.draft.DraftItem;
@@ -83,6 +82,7 @@ import ti4.service.unit.CheckUnitContainmentService;
 import ti4.service.user.AFKService;
 import ti4.settings.users.UserSettings;
 import ti4.settings.users.UserSettingsManager;
+import ti4.spring.jda.JdaService;
 
 public class Player extends PlayerProperties {
 
@@ -322,7 +322,7 @@ public class Player extends PlayerProperties {
     @JsonIgnore
     public Role getRoleForCommunity() {
         try {
-            return AsyncTI4DiscordBot.jda.getRoleById(getRoleIDForCommunity());
+            return JdaService.jda.getRoleById(getRoleIDForCommunity());
         } catch (Exception e) {
         }
         return null;
@@ -332,7 +332,7 @@ public class Player extends PlayerProperties {
     @JsonIgnore
     public MessageChannel getPrivateChannel() {
         try {
-            return AsyncTI4DiscordBot.jda.getTextChannelById(getPrivateChannelID());
+            return JdaService.jda.getTextChannelById(getPrivateChannelID());
         } catch (Exception e) {
             return null;
         }
@@ -1168,7 +1168,7 @@ public class Player extends PlayerProperties {
     @JsonIgnore
     public User getUser() {
         // TODO: This is to handle JDA being null during tests. We should think of a cleaner solution.
-        return AsyncTI4DiscordBot.jda == null ? null : AsyncTI4DiscordBot.jda.getUserById(getUserID());
+        return JdaService.jda == null ? null : JdaService.jda.getUserById(getUserID());
     }
 
     @Override
@@ -1176,7 +1176,7 @@ public class Player extends PlayerProperties {
         User userById = getUser();
         if (userById == null) return super.getUserName();
 
-        Member member = AsyncTI4DiscordBot.guildPrimary.getMemberById(getUserID());
+        Member member = JdaService.guildPrimary.getMemberById(getUserID());
         if (member == null) {
             setUserName(userById.getName());
         } else {
@@ -1251,7 +1251,7 @@ public class Player extends PlayerProperties {
             if (roleForCommunity == null && !getTeamMateIDs().isEmpty()) {
                 StringBuilder sb = new StringBuilder((noFactionIcon ? "" : getFactionEmoji()));
                 for (String userID : getTeamMateIDs()) {
-                    User userById = AsyncTI4DiscordBot.jda.getUserById(userID);
+                    User userById = JdaService.jda.getUserById(userID);
                     if (userById == null) {
                         continue;
                     }
