+++ src/main/java/ti4/message/logging/BotLogger.java
@@ -13,7 +13,6 @@ import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEve
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 import net.dv8tion.jda.api.exceptions.ErrorResponseException;
 import org.apache.commons.lang3.exception.ExceptionUtils;
-import ti4.AsyncTI4DiscordBot;
 import ti4.cron.CronManager;
 import ti4.helpers.DateTimeHelper;
 import ti4.helpers.DiscordWebhook;
@@ -22,6 +21,7 @@ import ti4.message.MessageHelper;
 import ti4.service.statistics.SREStats;
 import ti4.settings.GlobalSettings;
 import ti4.settings.GlobalSettings.ImplementedSettings;
+import ti4.spring.jda.JdaService;
 
 @UtilityClass
 public class BotLogger {
@@ -213,7 +213,7 @@ public class BotLogger {
                 if (channel == null) scheduleWebhookMessage(msgChunk); // Send message on webhook
                 else channel.sendMessage(msgChunk).queue(); // Send message on channel
             } else { // Handle error on last send
-                ThreadArchiveHelper.checkThreadLimitAndArchive(AsyncTI4DiscordBot.guildPrimary);
+                ThreadArchiveHelper.checkThreadLimitAndArchive(JdaService.guildPrimary);
 
                 if (channel == null) {
                     scheduleWebhookMessage(msgChunk); // Send message on webhook
@@ -340,7 +340,7 @@ public class BotLogger {
      */
     @Nullable
     private TextChannel getLogChannel(@Nonnull LogSeverity severity) {
-        Guild guild = AsyncTI4DiscordBot.guildPrimary;
+        Guild guild = JdaService.guildPrimary;
         if (guild == null) return null;
 
         return guild.getTextChannelsByName(severity.channelName, false).stream()
