+++ src/main/java/ti4/service/SusSlashCommandService.java
@@ -5,7 +5,6 @@ import lombok.experimental.UtilityClass;
 import net.dv8tion.jda.api.entities.channel.concrete.TextChannel;
 import net.dv8tion.jda.api.entities.channel.concrete.ThreadChannel;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
-import ti4.AsyncTI4DiscordBot;
 import ti4.helpers.Constants;
 import ti4.map.Game;
 import ti4.map.Player;
@@ -14,6 +13,7 @@ import ti4.map.persistence.ManagedGame;
 import ti4.message.MessageHelper;
 import ti4.service.fow.GMService;
 import ti4.service.game.GameNameService;
+import ti4.spring.jda.JdaService;
 
 @UtilityClass
 public class SusSlashCommandService {
@@ -110,7 +110,7 @@ public class SusSlashCommandService {
 
     public static void reportSusSlashCommand(SlashCommandInteractionEvent event, String jumpUrl) {
         TextChannel moderationLogChannel =
-                AsyncTI4DiscordBot.guildPrimary.getTextChannelsByName("moderation-log", true).stream()
+                JdaService.guildPrimary.getTextChannelsByName("moderation-log", true).stream()
                         .findFirst()
                         .orElse(null);
         if (moderationLogChannel == null) return;
