+++ src/main/java/ti4/service/async/DrumrollService.java
@@ -8,11 +8,11 @@ import java.util.function.Predicate;
 import lombok.experimental.UtilityClass;
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.channel.middleman.MessageChannel;
-import ti4.AsyncTI4DiscordBot;
 import ti4.map.Game;
 import ti4.map.persistence.GameManager;
 import ti4.message.MessageHelper;
 import ti4.service.emoji.MiscEmojis;
+import ti4.spring.jda.JdaService;
 
 @UtilityClass
 public class DrumrollService {
@@ -48,7 +48,7 @@ public class DrumrollService {
                 }
                 iteration++;
                 sleepForTwoSeconds();
-                if (!AsyncTI4DiscordBot.isReadyToReceiveCommands())
+                if (!JdaService.isReadyToReceiveCommands())
                     break; // if the bot needs to shut down, cancel all drumrolls
             }
             msg.delete().queue(null, null);
