+++ src/main/java/ti4/spring/resilience/CircuitBreakerFilter.java
@@ -12,8 +12,8 @@ import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
-import ti4.AsyncTI4DiscordBot;
 import ti4.executors.CircuitBreaker;
+import ti4.spring.jda.JdaService;
 
 @Order(Ordered.HIGHEST_PRECEDENCE)
 @RequiredArgsConstructor
@@ -33,7 +33,7 @@ public class CircuitBreakerFilter extends OncePerRequestFilter {
                     HttpStatus.SERVICE_UNAVAILABLE.value(), SERVICE_UNAVAILABLE_MESSAGE + "circuit breaker is open");
             return;
         }
-        if (!AsyncTI4DiscordBot.isReadyToReceiveCommands()) {
+        if (!JdaService.isReadyToReceiveCommands()) {
             response.sendError(
                     HttpStatus.SERVICE_UNAVAILABLE.value(), SERVICE_UNAVAILABLE_MESSAGE + "bot is not ready");
             return;
